students/k3340/laboratory_works/Bazhenov_Alexey/lab3-4/requirements.txt
django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
psycopg[binary]==3.2.13
python-dotenv==1.0.1
django-environ==0.11.2
django-cors-headers==4.9.0
